Split the tests
