+++ kafka_consumer/tests/python_client/__init__.py
@@ -1,0 +1,3 @@
+# (C) Datadog, Inc. 2023-present
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
