+++ kafka_consumer/tests/python_client/test_e2e.py
@@ -1,0 +1,13 @@
+# (C) Datadog, Inc. 2023-present
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+import pytest
+
+from ..common import assert_check_kafka
+
+pytestmark = [pytest.mark.e2e]
+
+
+def test_e2e(dd_agent_check, kafka_instance):
+    aggregator = dd_agent_check(kafka_instance)
+    assert_check_kafka(aggregator, kafka_instance['consumer_groups'])
