+++ kafka_consumer/tests/python_client/test_unit.py
@@ -1,4 +1,4 @@
-# (C) Datadog, Inc. 2018-present
+# (C) Datadog, Inc. 2023-present
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 import copy
@@ -11,14 +11,11 @@ from datadog_checks.base import ConfigurationError
 from datadog_checks.kafka_consumer import KafkaCheck
 from datadog_checks.kafka_consumer.client.kafka_python_client import OAuthTokenProvider
 
-from .common import BROKER_METRICS, CONSUMER_METRICS, KAFKA_CONNECT_STR
+from ..common import KAFKA_CONNECT_STR, metrics
 
-metrics = BROKER_METRICS + CONSUMER_METRICS
-
-pytestmark = [pytest.mark.integration]
+pytestmark = [pytest.mark.unit]
 
 
-@pytest.mark.unit
 def test_gssapi(kafka_instance, dd_run_check):
     instance = copy.deepcopy(kafka_instance)
     instance['sasl_mechanism'] = 'GSSAPI'
@@ -31,7 +28,6 @@ def test_gssapi(kafka_instance, dd_run_check):
         dd_run_check(kafka_consumer_check)
 
 
-@pytest.mark.unit
 def test_tls_config_ok(kafka_instance_tls):
     with mock.patch('datadog_checks.base.utils.tls.ssl') as ssl:
         with mock.patch('kafka.KafkaAdminClient') as kafka_admin_client:
@@ -96,7 +92,6 @@ def test_oauth_config(sasl_oauth_token_provider, expected_exception):
         check.check(instance)
 
 
-@pytest.mark.unit
 def test_oauth_token_client_config(kafka_instance):
     instance = copy.deepcopy(kafka_instance)
     instance['kafka_client_api_version'] = "3.3.2"
@@ -140,103 +135,6 @@ def test_tls_config_legacy(extra_config, expected_http_kwargs, kafka_instance):
         k: v for k, v in kafka_consumer_check._tls_context_wrapper.config.items() if k in expected_http_kwargs
     }
     assert expected_http_kwargs == actual_options
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_check_kafka(aggregator, kafka_instance, dd_run_check):
-    """
-    Testing Kafka_consumer check.
-    """
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [kafka_instance])
-    dd_run_check(kafka_consumer_check)
-    assert_check_kafka(aggregator, kafka_instance['consumer_groups'])
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_can_send_event(aggregator, kafka_instance, dd_run_check):
-    """
-    Testing Kafka_consumer check.
-    """
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [kafka_instance])
-    kafka_consumer_check.send_event("test", "test", [], "test", "test")
-    aggregator.assert_event("test", exact_match=False, count=1)
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_check_kafka_metrics_limit(aggregator, kafka_instance, dd_run_check):
-    """
-    Testing Kafka_consumer check.
-    """
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {'max_partition_contexts': 1}, [kafka_instance])
-    dd_run_check(kafka_consumer_check)
-
-    assert len(aggregator._metrics) == 1
-
-
-@pytest.mark.e2e
-def test_e2e(dd_agent_check, kafka_instance):
-    aggregator = dd_agent_check(kafka_instance)
-    assert_check_kafka(aggregator, kafka_instance['consumer_groups'])
-
-
-def assert_check_kafka(aggregator, consumer_groups):
-    for name, consumer_group in consumer_groups.items():
-        for topic, partitions in consumer_group.items():
-            for partition in partitions:
-                tags = [f"topic:{topic}", f"partition:{partition}"] + ['optional:tag1']
-                for mname in BROKER_METRICS:
-                    aggregator.assert_metric(mname, tags=tags, at_least=1)
-                for mname in CONSUMER_METRICS:
-                    aggregator.assert_metric(
-                        mname,
-                        tags=tags + [f"consumer_group:{name}"],
-                        at_least=1,
-                    )
-
-    aggregator.assert_all_metrics_covered()
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_consumer_config_error(caplog, dd_run_check):
-    instance = {'kafka_connect_str': KAFKA_CONNECT_STR, 'tags': ['optional:tag1']}
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [instance])
-
-    dd_run_check(kafka_consumer_check, extract_message=True)
-    assert 'monitor_unlisted_consumer_groups is False' in caplog.text
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_no_topics(aggregator, kafka_instance, dd_run_check):
-    kafka_instance['consumer_groups'] = {'my_consumer': {}}
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [kafka_instance])
-    dd_run_check(kafka_consumer_check)
-
-    assert_check_kafka(aggregator, {'my_consumer': {'marvel': [0]}})
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_no_partitions(aggregator, kafka_instance, dd_run_check):
-    kafka_instance['consumer_groups'] = {'my_consumer': {'marvel': []}}
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [kafka_instance])
-    dd_run_check(kafka_consumer_check)
-
-    assert_check_kafka(aggregator, {'my_consumer': {'marvel': [0]}})
-
-
-@pytest.mark.usefixtures('dd_environment')
-def test_version_metadata(datadog_agent, kafka_instance, dd_run_check):
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [kafka_instance])
-    kafka_consumer_check.check_id = 'test:123'
-
-    kafka_client = kafka_consumer_check.client.create_kafka_admin_client()
-    version_data = [str(part) for part in kafka_client._client.check_version()]
-    kafka_client.close()
-    version_parts = {f'version.{name}': part for name, part in zip(('major', 'minor', 'patch'), version_data)}
-    version_parts['version.scheme'] = 'semver'
-    version_parts['version.raw'] = '.'.join(version_data)
-
-    dd_run_check(kafka_consumer_check)
-    datadog_agent.assert_metadata('test:123', version_parts)
 
 
 # Get rid of dd_environment when we refactor and fix expected behaviors in the check
@@ -317,27 +215,3 @@ def test_config(dd_run_check, instance, aggregator, expected_exception, metric_c
 
     for m in metrics:
         aggregator.assert_metric(m, count=metric_count)
-
-
-@pytest.mark.parametrize(
-    'is_enabled, metric_count, topic_tags',
-    [
-        pytest.param(True, 4, ['topic:marvel', 'topic:dc'], id="Enabled"),
-        pytest.param(False, 2, ['topic:marvel'], id="Disabled"),
-    ],
-)
-@pytest.mark.usefixtures('dd_environment')
-def test_monitor_broker_highwatermarks(dd_run_check, aggregator, is_enabled, metric_count, topic_tags):
-    instance = {
-        'kafka_connect_str': KAFKA_CONNECT_STR,
-        'consumer_groups': {'my_consumer': {'marvel': None}},
-        'monitor_all_broker_highwatermarks': is_enabled,
-    }
-    check = KafkaCheck('kafka_consumer', {}, [instance])
-    dd_run_check(check)
-
-    # After refactor and library migration, write unit tests to assert expected metric values
-    aggregator.assert_metric('kafka.broker_offset', count=metric_count)
-    for tag in topic_tags:
-        aggregator.assert_metric_has_tag('kafka.broker_offset', tag, count=2)
-    aggregator.assert_metric_has_tag_prefix('kafka.broker_offset', 'partition', count=metric_count)
